Add back the use of INSTALL_SCRIPT correctly
Due to a momentary lapse of reason, the changes of 52fa489337 (Install
wx-config as a script, not as a binary program, 2018-08-16) were applied
to the generated Makefile.in file and so were promptly lost during the
next rebake.

Reapply them properly now, by modifying the .bkl file from which this
file is generated.

Note that rebaking now requires bakefile v0.2.12, which adds support for
defining INSTALL_SCRIPT.

See #18197.